navtex: sync using multicorrelator, instead of null crossings
Null crossings are somewhat noisy, and the code to keep the navtex
decoder in sync with the incoming signal using null crossings was
rather fragile.

Use a multicorrelator instead, which relies on the averaged magnitude
of the signal accumulator to sync the decoder with the incoming signal.

When debugging the code, the multicorrelator mostly corrects the
modem forward in time, which can be explained by the fact that a
bit takes 110.25 samples, while the code uses 110. When the NAVTEX
transmitter is running at exactly 100 baud, one can expect to see
the decoder get adjusted 25 times a second, to make up for the
difference between 11000 and 11025.

However, I have also seen it recenter the modem earlier on occasion.

When multiple signals are on the air simultaneously, the null crossing
code would often lose track of the signal. The multicorrelator seems
to be more stable in this situation, though of course when both signals
are close in strength things do not get decoded right.

The signal sampling spread of 1/6 of the width of a bit was set through
trial and error. A larger spread increases the signal difference between
early, prompt, and late samples, but reduces the accumulator value seen
by the demodulator. A smaller spread increases the accumulator value seen,
but makes it harder to lock on in noisy conditions.

Signed-off-by: Rik van Riel <riel@surriel.com>