[PATCH] display navtex S/N ratio
The NAVTEX display_metric() function was buggy, in that decayavg
returns the decayed value, but does not store it. It always put
the current value in the metric, and kept avg_ratio at 0.0.

This resulted in a somewhat chaotic, and not very useful metric
display. Copy over the S/N calculation from the RTTY code, because
that code seems to work well.

Also print the S/N in Status2, like the RTTY code and other modes
do.

I am not convinced the RTTY S/N code works correctly on NAVTEX
signals. The wave form appears to be different, with much more
energy in the middle between mark and space.

Signed-off-by: Rik van Riel <riel@surriel.com>