navtex: widen afc filter for 'jump 90 Hz' code
When the NAVTEX code spots a power imbalance of more than a factor
5 between mark and space, it will shift the frequency by 90 Hz.
This is reported to help with some signals.

However, it breaks with some other signals, which have a different
spectral distribution between mark and space, with a spectrum looking
something like this:

                                       *
                                       *
                                       *
                                      **
    ******                           ***
   ********                         ******
  **********                       ********
 ********************************************
**********************************************

In this spectrum, mark & space have a similar amount of energy,
but that is only apparent when the comparison between them is
done on a wider sample than 10 Hz.

Sampling 30 Hz instead seems to result in a more stable AFC.

Signed-off-by: Rik van Riel <riel@surriel.com>