navtex: make early/prompt/late detector able to lock onto more signals
When the early/prompt/late detector was totally out of phase, it sometimes
totally failed to lock onto signals.

Fix that by making the detectors jump the entire distance between them if
the prompt detector is at a lower average than both the early and late
detectors, and limit that to one jump by copying over the average values.

Signed-off-by: Rik van Riel <riel@surriel.com>